testing/web-platform/tests/old-tests/submission/Infraware/Session_History/contents/Session_History/location_port.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>location_port</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		
		if (('location' in window) && window.location != null
			&& ('location' in document) && document.location != null){
			var port = location.port;
			/*var url = location.href;
			
			var pos = url.indexOf("//");
			if (pos != -1) {
				url = url.substr(pos+2, url.length-pos-2);
				pos = url.indexOf(":");
				if (pos != -1) {
					var pos1 = url.indexOf("/");
					if (pos != -1)
						url = url.substr(pos+1, pos1-pos-1);
				}
			}
			
			test(function() {
				assert_equals(port, url, "port attribute is not correct.");
			});*/
			test(function() {
				assert_equals(port, "", "port attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("location interface is not supported.");
			});
		}

	</script>
 </body>
</html>
